+++ src/main/java/io/appium/java_client/android/AndroidMobileCommandHelper.java
@@ -64,6 +64,18 @@ public class AndroidMobileCommandHelper extends MobileCommand {
     }
 
     /**
+     * This method forms a {@link java.util.Map} of parameters to
+     * Retrieve the display density of the Android device.
+     *
+     * @return a key-value pair. The key is the command name. The value is a
+     * {@link java.util.Map} command arguments.
+     */
+    public static Map.Entry<String, Map<String, ?>> getDisplayDensityCommand() {
+        return new AbstractMap.SimpleEntry<>(
+            GET_DISPLAY_DENSITY, ImmutableMap.<String, Object>of());
+    }
+
+    /**
      * This method forms a {@link java.util.Map} of parameters for the
      * getting of a network connection value.
      *
@@ -73,6 +85,18 @@ public class AndroidMobileCommandHelper extends MobileCommand {
     public static Map.Entry<String, Map<String, ?>> getNetworkConnectionCommand() {
         return new AbstractMap.SimpleEntry<>(
                 GET_NETWORK_CONNECTION, ImmutableMap.<String, Object>of());
+    }
+
+    /**
+     * This method forms a {@link java.util.Map} of parameters to
+     * Retrieve visibility and bounds information of the status and navigation bars.
+     *
+     * @return a key-value pair. The key is the command name. The value is a
+     * {@link java.util.Map} command arguments.
+     */
+    public static Map.Entry<String, Map<String, ?>> getSystemBarsCommand() {
+        return new AbstractMap.SimpleEntry<>(
+            GET_SYSTEM_BARS, ImmutableMap.<String, Object>of());
     }
 
     /**
