+++ src/test/java/io/appium/java_client/android/AndroidDriverTest.java
@@ -134,4 +134,10 @@ public class AndroidDriverTest extends BaseAndroidTest {
         Map<?,?> map = (Map<?, ?>) driver.getSessionDetail("desired");
         assertNotEquals(map.size(), 0);
     }
+    
+    @Test public void deviceDetailsAndKeyboardTest() {
+        assertFalse(driver.isKeyboardShown());
+        assertNotNull(driver.getDisplayDensity());
+        assertNotEquals(0, driver.getSystemBars().size());
+    }
 }
